requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
anyio==4.2.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
babel==2.16.0
beautifulsoup4==4.12.2
billiard==4.2.0
black==23.7.0
boto3==1.34.113
botocore==1.34.113
Brotli==1.1.0
cachetools==5.5.0
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
channels==3.0.4
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.2
cssselect2==0.7.0
currency.converter==0.5.5
daphne==3.0.2
defusedxml==0.7.1
Django==4.1.4
django-appconf==1.0.5
django-celery-beat==2.7.0
django-compressor==4.4
django-cors-headers==4.2.0
django-redis==5.4.0
django-sass-processor==1.2.2
django-single-session==0.2.0
django-storages==1.14.3
django-timezone-field==7.0
djangorestframework==3.14.0
docx2txt==0.8
et-xmlfile==1.1.0
exceptiongroup==1.2.0
filelock==3.16.1
filetype==1.2.0
fonttools==4.45.1
forex-python==1.8
frozenlist==1.4.0
fsspec==2024.10.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.26.2
hyperlink==21.0.0
ibge-api==0.1.1
idna==3.4
incremental==22.10.0
jmespath==1.0.1
joblib==1.4.2
kombu==5.3.6
libsass==0.22.0
lxml==5.3.1
msgpack==1.0.8
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
numpy==1.26.3
oauthlib==3.2.2
openai==0.28.1
opencv-python==4.11.0.86
openpyxl==3.1.5
oscrypto==1.3.0
packaging==23.1
pandas==2.2.0
pathspec==0.11.2
pika==1.3.2
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.43
propcache==0.2.1
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==22.3.5
pycparser==2.21
pydyf==0.8.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.7
PyJWT==2.8.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyOpenSSL==24.1.0
pypdf==5.3.1
PyPDF2==3.0.1
pyphen==0.14.0
pypng==0.20220715.0
pytesseract==0.3.10
python-bidi==0.6.6
python-crontab==3.2.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.2
qrcode==7.4.2
rcssmin==1.1.1
redis==5.0.3
regex==2024.11.6
reportlab==4.3.1
requests==2.31.0
requests-oauthlib==2.0.0
rjsmin==1.2.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
service-identity==24.1.0
setuptools==75.8.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.20.3
tomli==2.0.1
tqdm==4.66.1
transformers==4.46.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.3.1
Unidecode==1.3.6
uritools==4.0.3
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
weasyprint==60.1
webencodings==0.5.1
xhtml2pdf==0.2.17
yarl==1.9.2
zope.interface==6.3
zopfli==0.2.3
